Move out valueBasedDecimalFormatter to separate file
